@@ -7,34 +7,34 @@ using SQBuilder.Enums;
 namespace SQBuilder
 {
     public abstract partial class ScriptBuilder : IScriptBuilder
-	{
-		public string Schema { get; private set;}
-		public string InicialDelimiter { get; private set;}
-		public string FinalDelimiter { get; private set;}
+    {
+        public string Schema { get; private set; }
+        public string InicialDelimiter { get; private set; }
+        public string FinalDelimiter { get; private set; }
 
-		protected string _query { get; set; }
-		protected List<string> _queries { get; set; }
-		protected List<string> _select { get; set; }
-		protected List<string> _from { get; set; }
-		protected List<string> _where { get; set; }
-		protected List<Tuple<EJoinTypes, string>> _join { get; set; }
-		protected List<string> _groupBy { get; set; }
-		protected List<string> _orderBy { get; set; }
-		protected List<string> _having { get; set; }
-		protected uint _top { get; set; }
-		public ScriptBuilder()
-		{
+        protected string _query { get; set; }
+        protected List<string> _queries { get; set; }
+        protected List<string> _select { get; set; }
+        protected List<string> _from { get; set; }
+        protected List<string> _where { get; set; }
+        protected List<Tuple<EJoinTypes, string>> _join { get; set; }
+        protected List<string> _groupBy { get; set; }
+        protected List<string> _orderBy { get; set; }
+        protected List<string> _having { get; set; }
+        protected uint _top { get; set; }
+        public ScriptBuilder()
+        {
             InitiateLists();
             _queries = new List<string>();
         }
 
-		protected void SetTop(uint top)
-		{
-			if (top > 0) _top = top;
-		}
+        protected void SetTop(uint top)
+        {
+            if (top > 0) _top = top;
+        }
 
-		public void InitiateLists()
-		{
+        public void InitiateLists()
+        {
             _select = new List<string>();
             _from = new List<string>();
             _where = new List<string>();
@@ -44,83 +44,83 @@ namespace SQBuilder
             _having = new List<string>();
         }
 
-		#region Tranformar em uma consulta
-		public virtual string ToScript()
-		{
-			_query = string.Empty;
-
-			AssemblyUnion();
-			AssemblySelect();
-			AssemblyFrom();
-			AssemblyJoin();
-			AssemblyWhere();
-			AssemblyGroupBy();
-			AssemblyHaving();
-			AssemblyOrderBy();
-
-			return Regex.Replace(_query, @"\s+", " ", RegexOptions.Multiline).Trim();
-		}
-		#endregion
-
-		public virtual void AssemblyUnion()
+        #region Tranformar em uma consulta
+        public virtual string ToScript()
         {
-			if (_queries.Count > 0)
-            _query += string.Join(@" ", _queries.Where(d => !string.IsNullOrWhiteSpace(d)));
-		}
+            _query = string.Empty;
 
-		public virtual void AssemblySelect()
+            AssemblyUnion();
+            AssemblySelect();
+            AssemblyFrom();
+            AssemblyJoin();
+            AssemblyWhere();
+            AssemblyGroupBy();
+            AssemblyHaving();
+            AssemblyOrderBy();
+
+            return Regex.Replace(_query, @"\s+", " ", RegexOptions.Multiline).Trim();
+        }
+        #endregion
+
+        public virtual void AssemblyUnion()
         {
-			if (_select.Count > 0)
-				_query += " SELECT ";
+            if (_queries.Count > 0)
+                _query += string.Join(@" ", _queries.Where(d => !string.IsNullOrWhiteSpace(d)));
+        }
 
-			_query += string.Join(@", ", _select.Where(d => !string.IsNullOrWhiteSpace(d)));
-		}
-
-		public virtual void AssemblyFrom()
-		{
-			if (_select.Count > 0)
-				_query += " FROM ";
-
-			_query += string.Join(@", ", _from.Where(d => !string.IsNullOrWhiteSpace(d)));
-		}
-
-		public virtual void AssemblyJoin()
+        public virtual void AssemblySelect()
         {
-			if (_join.Count > 0)
-				foreach (Tuple<EJoinTypes, string> item in _join)
-					_query += $"{item} ";
-		}
+            if (_select.Count > 0)
+                _query += " SELECT ";
 
-		public virtual void AssemblyWhere()
+            _query += string.Join(@", ", _select.Where(d => !string.IsNullOrWhiteSpace(d)));
+        }
+
+        public virtual void AssemblyFrom()
         {
-			if (_where.Count > 0)
-				_query += " WHERE ";
+            if (_select.Count > 0)
+                _query += " FROM ";
 
-			_query += string.Join(@" AND ", _where.Where(d => !string.IsNullOrWhiteSpace(d)));
-		}
+            _query += string.Join(@", ", _from.Where(d => !string.IsNullOrWhiteSpace(d)));
+        }
 
-		public virtual void AssemblyGroupBy()
-		{
-			if (_groupBy.Count > 0)
-				_query += " GROUP BY ";
+        public virtual void AssemblyJoin()
+        {
+            if (_join.Count > 0)
+                foreach (Tuple<EJoinTypes, string> item in _join)
+                    _query += $"{item} ";
+        }
 
-			_query += string.Join(@", ", _groupBy.Where(d => !string.IsNullOrWhiteSpace(d)));
-		}
+        public virtual void AssemblyWhere()
+        {
+            if (_where.Count > 0)
+                _query += " WHERE ";
 
-		public virtual void AssemblyOrderBy()
-		{
-			if (_orderBy.Count > 0)
-				_query += " ORDER BY ";
+            _query += string.Join(@" AND ", _where.Where(d => !string.IsNullOrWhiteSpace(d)));
+        }
 
-			_query += string.Join(@", ", _orderBy.Where(d => !string.IsNullOrWhiteSpace(d)));
-		}
+        public virtual void AssemblyGroupBy()
+        {
+            if (_groupBy.Count > 0)
+                _query += " GROUP BY ";
 
-		public virtual void AssemblyHaving()
-		{
-			if (_having.Count > 0)
-				_query += " HAVING ";
+            _query += string.Join(@", ", _groupBy.Where(d => !string.IsNullOrWhiteSpace(d)));
+        }
 
-			_query += string.Join(@" AND ", _having.Where(d => !string.IsNullOrWhiteSpace(d)));
-		}
+        public virtual void AssemblyOrderBy()
+        {
+            if (_orderBy.Count > 0)
+                _query += " ORDER BY ";
+
+            _query += string.Join(@", ", _orderBy.Where(d => !string.IsNullOrWhiteSpace(d)));
+        }
+
+        public virtual void AssemblyHaving()
+        {
+            if (_having.Count > 0)
+                _query += " HAVING ";
+
+            _query += string.Join(@" AND ", _having.Where(d => !string.IsNullOrWhiteSpace(d)));
+        }
     }
 }
