@@ -5,16 +5,16 @@ namespace SQBuilder
 {
     public abstract partial class ScriptBuilder : IScriptBuilder
     {
-		/// <summary>
-		/// Adicione apenas o nome da tabela e referência ON, a classe adiciona a instrução INNER JOIN
-		/// </summary>
-		/// <param name="content"></param>
-		public virtual IScriptBuilder InnerJoin(string content)
-		{
-			if (!string.IsNullOrWhiteSpace(content))
-				_join.AddContent(Tuple.Create(EJoinTypes.InnerJoin, content));
-			return this;
-		}
+        /// <summary>
+        /// Adicione apenas o nome da tabela e referência ON, a classe adiciona a instrução INNER JOIN
+        /// </summary>
+        /// <param name="content"></param>
+        public virtual IScriptBuilder InnerJoin(string content)
+        {
+            if (!string.IsNullOrWhiteSpace(content))
+                _join.AddContent(Tuple.Create(EJoinTypes.InnerJoin, content));
+            return this;
+        }
 
         /// <summary>
         /// Adicione apenas o nome da tabela e referência ON, a classe adiciona a instrução LEFT JOIN
@@ -25,6 +25,13 @@ namespace SQBuilder
             string clausure = $"{content} ON {on}";
 
             return InnerJoin(clausure);
+        }
+
+        //TODO
+        public virtual IScriptBuilder InnerJoin<TModel>(string alias)
+        {
+
+            return InnerJoin("");
         }
     }
 }
